+++ ivp/ivp_collision/ivp_compact_ledge_solver.cxx
@@ -698,16 +698,16 @@ IVP_BOOL IVP_Compact_Ledge_Solver::check_ledge(const IVP_Compact_Ledge *cl)
     // For debugging only
     // Checks for consistency
 
-	IVP_ASSERT ( (int(cl) & 15) == 0);
+	IVP_ASSERT ( (intptr_t(cl) & 15) == 0);
 	const IVP_Compact_Poly_Point *ppppp=cl->get_point_array();
-	IVP_ASSERT ( ((int)ppppp & 15 ) == 0);
+	IVP_ASSERT ( ((intptr_t)ppppp & 15 ) == 0);
 
     // all triangles 
     const IVP_Compact_Triangle *tri = cl->get_first_triangle();
     for(int i=0; i<cl->get_n_triangles(); i++){
 	IVP_ASSERT(i == tri->get_tri_index());
 
-		IVP_ASSERT ( (int(tri) & 15) == 0);
+		IVP_ASSERT ( (intptr_t(tri) & 15) == 0);
 
 	// all edges
 	for(int j=0; j<3; j++){
