+++ ivp/ivp_compact_builder/ivp_compact_recursive.cxx
@@ -103,7 +103,10 @@ void IVP_Compact_Recursive::set_rekursive_convex_hull(){
       int pi[3];
       // convert points to unique point nums
       for (int i = 0; i<3;i++){
-	int ind = (int)point_hash.find( (char *)p[i]);
+    //lwss - x64 fixes
+	//int ind = (int)point_hash.find( (char *)p[i]);
+	intptr_t ind = (intptr_t)point_hash.find( (char *)p[i]);
+	//lwss end
 	if (ind<0){
 	  point_hash.add( (char *)p[i], (void *)n_points_in_hash);
 	  ind = n_points_in_hash++;
@@ -140,7 +143,10 @@ void IVP_Compact_Recursive::set_rekursive_convex_hull(){
       int pi[3];
       // convert points to unique point nums
       for (int i = 0; i<3;i++){
-	int ind = (int)point_hash.find( (char *)p[i]);
+    //lwss -x64 fixes
+	//int ind = (int)point_hash.find( (char *)p[i]);
+	intptr_t ind = (intptr_t)point_hash.find( (char *)p[i]);
+	//lwss end
 	pi[i] = ind;
       }
       // find triangle in hash
