+++ ivp/ivp_compact_builder/ivp_surbuild_pointsoup.cxx
@@ -357,7 +357,10 @@ IVP_Compact_Ledge *IVP_SurfaceBuilder_Pointsoup::try_to_build_convex_ledge_from_
 	      int max_index = 0;
 	      int max_index2 = 0;
 	      for (int i0 = 0; i0 < point_indizes.len(); i0++){
-		int in = int(point_indizes.element_at(i0));
+        //lwss - x64 fixes
+        //int in = int(point_indizes.element_at(i0));
+        intptr_t in = intptr_t(point_indizes.element_at(i0));
+		//lwss end
 		if (skip_list[in]) goto no_point_skipped;
 		IVP_U_Point *p2 = plane->points.element_at( i0 );
 		IVP_DOUBLE dist = plane->points.element_at( 0 )->quad_distance_to(p2);
@@ -383,7 +386,10 @@ IVP_Compact_Ledge *IVP_SurfaceBuilder_Pointsoup::try_to_build_convex_ledge_from_
 	      {
 		  for (int i2 = 0; i2 < point_indizes.len(); i2++){
 		    if (i2 != max_index2 && i2 != max_index) {
-		      int in = int(point_indizes.element_at(i2));
+              //lwss - x64 fixes
+		      //int in = int(point_indizes.element_at(i2));
+		      intptr_t in = intptr_t(point_indizes.element_at(i2));
+		      //lwss end
 		      skip_list[in]++;
 		      IVP_ASSERT( use_list[in] );
 		      //printf("point removed %i %i\n",in, points->len());
