+++ ivp/ivp_surface_manager/ivp_compact_grid.hxx
@@ -14,6 +14,7 @@
 #ifndef _IVP_COMPACT_GRID_INCLUDED
 #define _IVP_COMPACT_GRID_INCLUDED
 
+#include <cstdint> //lwss - x64 fixes
 
 /********************************************************************************
  *  Class:	    IVP_Compact_Grid_Element
@@ -67,7 +68,9 @@ public:
      *  Description:    INTERNAL METHOD
      *****************************************************************************/
     const IVP_Compact_Grid_Element *get_grid_elements() {
-    	char *base = (char *)(((int)this) + this->offset_grid_elements);
+        //lwss - x64 fixes
+    	char *base = (char *)(((intptr_t)this) + this->offset_grid_elements);
+    	//lwss end
 	return((const IVP_Compact_Grid_Element *)base);
     }
 
@@ -76,7 +79,7 @@ public:
      *  Description:    INTERNAL METHOD
      *****************************************************************************/
     const IVP_Compact_Ledge *get_compact_ledge_at(int i) {
-    	char *base = (char *)(((int)this) + this->offset_compact_ledge_array[i]);
+    	char *base = (char *)(((intptr_t)this) + this->offset_compact_ledge_array[i]);
 	return((const IVP_Compact_Ledge *)base);
     }
